+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Buffs/ conditions are toggleable from the statusHUD (#329)
 - Fix max dex bonus being calculated wrong sometimes (thanks mkahvi!) (#490)
 - SR notes were not listed in the defenses summary when the actor had no SR
+- Fix charge linking using Items from the equipment category
 
 ### Changelog